requirements.txt
# Système de Révision Intelligent - Dépendances

# IA et NLP
google-generativeai>=0.8.0
langchain>=0.1.0
langchain-google-genai>=2.0.0

# Traitement de documents
PyPDF2>=3.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
markdown>=3.5.0

# Extraction de texte
pdfplumber>=0.10.0
pytesseract>=0.3.10  # Pour OCR si nécessaire

# Manipulation de données
pandas>=2.0.0
numpy>=1.24.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Interface utilisateur
streamlit>=1.28.0
rich>=13.0.0          # Pour une belle interface terminal
questionary>=2.0.0    # Pour les menus interactifs

# Visualisation
matplotlib>=3.7.0
networkx>=3.0         # Pour visualiser les liens entre concepts

# Calendrier et planning
icalendar>=5.0.0
python-dateutil>=2.8.0

# Base de données locale
tinydb>=4.8.0         # Base de données légère JSON
